refactor(dashboard): compose plan rows from Item directly
Each row writes its own Item with the shared column template instead of going
through a wrapper that only forwarded props. plan-row.tsx keeps the parts that
must agree between rows: the columns, the plan and price cells, and the action.
